chore(wire): regenerate TS types for FsBrowse/FsEntry
Server commit a09e5ac added FsEntryResponse/FsBrowseResponse to
RESPONSE_MODELS but the generated TS wasn't refreshed; wire-drift
check caught it. Dedup the hand-written FsEntry/FsBrowseResponse in
types.ts by re-exporting from the generated module.
